reuse expand_wildcard in logical plan builder
